+++ src/magpylib_jax/magnet/cylinder.py
@@ -28,7 +28,6 @@ class Cylinder(BaseSource):
         self.dimension = dimension
         self.polarization = polarization
         self.magnetization = magnetization
-        self.meshing = None
         super().__init__(
             position=position,
             orientation=orientation,